refactor: extract scale extent validation
