+++ svg-time-series/src/chart/zoomState.ts
@@ -3,6 +3,7 @@ import { zoom as d3zoom, zoomIdentity, zoomTransform } from "d3-zoom";
 import type { D3ZoomEvent, ZoomBehavior, ZoomTransform } from "d3-zoom";
 import { ZoomScheduler, sameTransform } from "./zoomScheduler.ts";
 import type { RenderState } from "./render.ts";
+import { validateScaleExtent } from "./zoomUtils.ts";
 
 export { sameTransform };
 
@@ -33,38 +34,6 @@ export class ZoomState {
     return node && this.zoomAreaNode ? node : null;
   }
 
-  public static validateScaleExtent(extent: unknown): [number, number] {
-    const error = () =>
-      new Error(
-        `scaleExtent must be two finite, positive numbers where extent[0] < extent[1]. Received: ${Array.isArray(extent) ? `[${extent.join(",")}]` : String(extent)}`,
-      );
-
-    if (!Array.isArray(extent) || extent.length !== 2) {
-      throw error();
-    }
-
-    const [min, max] = extent as [unknown, unknown];
-
-    if (
-      typeof min !== "number" ||
-      typeof max !== "number" ||
-      !Number.isFinite(min) ||
-      !Number.isFinite(max)
-    ) {
-      throw error();
-    }
-
-    if (min <= 0 || max <= 0) {
-      throw error();
-    }
-
-    if (min >= max) {
-      throw error();
-    }
-
-    return [min, max];
-  }
-
   constructor(
     private zoomArea: Selection<SVGRectElement, unknown, HTMLElement, unknown>,
     private state: RenderState,
@@ -74,7 +43,7 @@ export class ZoomState {
     ) => void = () => {},
     options: IZoomStateOptions = { scaleExtent: [1, 40] },
   ) {
-    this.scaleExtent = ZoomState.validateScaleExtent(options.scaleExtent);
+    this.scaleExtent = validateScaleExtent(options.scaleExtent);
     this.zoomBehavior = d3zoom<SVGRectElement, unknown>()
       .scaleExtent(this.scaleExtent)
       .translateExtent([
@@ -115,7 +84,7 @@ export class ZoomState {
     if (!node) {
       return;
     }
-    this.scaleExtent = ZoomState.validateScaleExtent(extent);
+    this.scaleExtent = validateScaleExtent(extent);
     this.zoomBehavior.scaleExtent(this.scaleExtent);
     const current = zoomTransform(node);
     const [min, max] = this.scaleExtent;
